programmatic way to create guide
a table of contents list contains all of the sections for a particular OS. the
create script will search the /res directory for <section>-<label>.html files.
if a section is not yet available, a TODO is added. this should make it easier
to manage sections and their contents among OS versions.