require specific version override
previously, was able to add -<version> file in /res where <version> matched
the short name of the operating system version. this makes it easier to add
content by not having to edit the respective /toc file. however, explicit
overriding to specify version is better practice. it also allows comparison
of the /toc files to see where there are differences between operating system
versions.